# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SIKURI SI - Sistem Informasi Kurikulum Program Studi S1 Sistem Informasi
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profil-lulusan:
    get:
      operationId: listProfilLulusan
      tags: [profil-lulusan]
      summary: Daftar Profil Lulusan
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfilLulusan"
    post:
      operationId: createProfilLulusan
      tags: [profil-lulusan]
      summary: Tambah Profil Lulusan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfilLulusanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfilLulusan"
  /profil-lulusan/{id}:
    patch:
      operationId: updateProfilLulusan
      tags: [profil-lulusan]
      summary: Update Profil Lulusan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfilLulusanUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfilLulusan"
    delete:
      operationId: deleteProfilLulusan
      tags: [profil-lulusan]
      summary: Hapus Profil Lulusan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── BAHAN KAJIAN ─────────────────────────────────────────────────────────────
  /bahan-kajian:
    get:
      operationId: listBahanKajian
      tags: [bahan-kajian]
      summary: Daftar Bahan Kajian
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BahanKajian"
    post:
      operationId: createBahanKajian
      tags: [bahan-kajian]
      summary: Tambah Bahan Kajian
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BahanKajianInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BahanKajian"
  /bahan-kajian/{id}:
    patch:
      operationId: updateBahanKajian
      tags: [bahan-kajian]
      summary: Update Bahan Kajian
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BahanKajianUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BahanKajian"
    delete:
      operationId: deleteBahanKajian
      tags: [bahan-kajian]
      summary: Hapus Bahan Kajian
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CPL SN-DIKTI ─────────────────────────────────────────────────────────────
...
  /pemetaan/cpl-mk/toggle:
    post:
      operationId: togglePemetaanCplMk
      tags: [pemetaan]
      summary: Toggle relasi CPL-MK
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PemetaanToggle"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PemetaanToggleResult"
...
  /pemetaan/cpl-bk/toggle:
    post:
      operationId: togglePemetaanCplBk
      tags: [pemetaan]
      summary: Toggle relasi CPL-BK
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PemetaanToggle"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PemetaanToggleResult"
...
  /pemetaan/bk-mk/toggle:
    post:
      operationId: togglePemetaanBkMk
      tags: [pemetaan]
      summary: Toggle relasi BK-MK
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PemetaanToggle"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PemetaanToggleResult"
...
  /pemetaan/cpl-pl/toggle:
    post:
      operationId: togglePemetaanCplPl
      tags: [pemetaan]
      summary: Toggle relasi CPL-PL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PemetaanToggle"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PemetaanToggleResult"
  /pemetaan/mk-cpmk/toggle:
    post:
      operationId: togglePemetaanMkCpmk
      tags: [pemetaan]
      summary: Toggle relasi MK-CPMK
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PemetaanToggle"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PemetaanToggleResult"

  # ─── RPS ──────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ProfilLulusan:
      type: object
      required: [id, kode, deskripsi]
      properties:
        id:         { type: integer }
        kode:       { type: string }
        deskripsi:  { type: string }
        tipe:       { type: ["string", "null"] }
        referensi:  { type: ["string", "null"] }
        createdAt:  { type: string }
        updatedAt:  { type: string }
    ProfilLulusanInput:
      type: object
      required: [kode, deskripsi]
      properties:
        kode:      { type: string }
        deskripsi: { type: string }
        tipe:      { type: string }
        referensi: { type: string }
    ProfilLulusanUpdate:
      type: object
      properties:
        kode:      { type: string }
        deskripsi: { type: string }
        tipe:      { type: string }
        referensi: { type: string }

    # ── Bahan Kajian ───────────────────────────────────────────────────────────
    BahanKajian:
      type: object
      required: [id, kode, nama]
      properties:
        id:        { type: integer }
        kode:      { type: string }
        nama:      { type: string }
        createdAt: { type: string }
    BahanKajianInput:
      type: object
      required: [kode, nama]
      properties:
        kode: { type: string }
        nama: { type: string }
    BahanKajianUpdate:
      type: object
      properties:
        kode: { type: string }
        nama: { type: string }

    # ── CPL SN-Dikti ───────────────────────────────────────────────────────────
...
    PemetaanToggle:
      type: object
      required: [rowId, colId]
      properties:
        rowId: { type: integer }
        colId: { type: integer }
    PemetaanToggleResult:
      type: object
      required: [active]
      properties:
        active: { type: boolean }

    # ── RPS ────────────────────────────────────────────────────────────────────
...