# lib/api-zod/src/generated/types/pemetaanToggle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SIKURI SI - Sistem Informasi Kurikulum Program Studi S1 Sistem Informasi
 * OpenAPI spec version: 0.1.0
 */

export interface PemetaanToggle {
  rowId: number;
  colId: number;
}